Merge from mihon until mihon 0.16.2

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/Tracker.kt
@@ -24,9 +24,9 @@ interface Tracker {
     @ColorInt
     fun getLogoColor(): Int
 
-    fun getStatus(status: Int): StringResource?
+    fun getStatus(status: Long): StringResource?
 
-    fun getCompletionStatus(): Int
+    fun getCompletionStatus(): Long
 
     fun getScoreList(): ImmutableList<String>
